Add files via upload
Integrated Speech to Text to the Web Page
Important Note about the speech recognition: 
The way the audio recording works is that it waits for silence to indicate the end of a recording. There is no "stop recording" button. This is certainly something that we should try to look into adding/ fixing. For now though, just press the "Record" button, say whatever, then be quiet for maybe 3 seconds in order for the process to stop recording you then translate whatever to text.

To Get Everything to Work
------------------------------
To get everything to work you'll probably have to do some downloads and such. 
To get basic flask stuff going (the stuff from the last commit) 
-pip install flask
New wt form part
-pip install wtforms
If you're coding in pycharms you can input those commands into the terminal. The next one to get python to be able to do audio stuff you have to go to the following website, download something from there, then do the pip install
-go to https://www.lfd.uci.edu/~gohlke/pythonlibs/#pyaudio
-download the third option under the PyAudio section (it should be cp37 and have a amd64 in the extension)
-after downloading that go back to pycharm and use command "pip install PyAudio"
Finally to get the speech recognition part use the following command
-pip install SpeechRecognition

If the project doesn't work after doing those installs let me know and I can try to figure out if I forgot something. 
Cool note about the project so far. Without any extra work the translate page lets you click the record button again after already translating something to translate another sentence. 
I also made sure to include the link to the youtube video that I got the speech recognition code from at the top of speech2text.py. As mentioned at the top of these commit changes, there is some improvements that we might be able to make to the recording audio process so as time goes on speech2text.py will hopefully change and become more of our own. 
Other cool note is that in speech2text.py the line that says
audio = r.listen(source)
you should be able to change "source" to an audio file and instead of listening from your microphone the program will just listen to the inputted audio file. I never tested it, but utilizing this as an option that the user picks could be an interesting feature of our project.